Fix to module activation checkbox in module configuration pages
When the multishop option is activated and at least two shops have been created,
a new panel is displayed at the bottom of module configuration pages, containing a
checkbox labeled "Activate module for this shop context". Checking/unchecking
this checkbox is supposed to enable/disable the current module for the current shop
context.

This does not currently work. Clicking this box reloads the current url, adding
an "enable" parameter with a value of 0 or 1 depending on if it is checked or not. Getting
the status of the checkbox is achieved via the jquery .attr() method, which since jQuery 1.6
cannot be used for attributes like "checked" or "disabled". As a result "enable" is set to 0
wether the box is checked or not.

The solution is to use jQuery's .is(':checked') method instead.